requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyJWT==2.8.0
django-cors-headers==4.3.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1